Wrapped promise in promise
