+++ src/javascripts/components/cowList/cowList.js
@@ -1,17 +1,10 @@
 import './cowList.scss';
 import cowData from '../../helpers/data/cows';
-import { resolve } from 'url';
 
 const buildCows = () => {
   cowData.getCows()
     .then((response) => {
-      const demCows = response.data.cows;
-      const cows = [];
-      Object.keys(demCows).forEach((cowId) => {
-        demCows[cowId].id = cowId;
-        cows.push(demCows[cowId]);
-      });
-      console.error('it worked!', cows);
+      console.error('cow array from cowlist', response);
     })
     .catch((error) => {
       console.error('shit broke', error);
